+++ include/dynamic_gap/gap_feasibility/GapFeasibilityChecker.h
@@ -25,16 +25,18 @@ namespace dynamic_gap
             void updateEgoCircle(boost::shared_ptr<sensor_msgs::LaserScan const> msg_);
             
         private:
-            double gapSplinecheck(dynamic_gap::Gap & gap, dynamic_gap::Estimator*, dynamic_gap::Estimator*);
-            double generateCrossedGapTerminalPoints(double t, dynamic_gap::Gap & gap, 
-                                                    dynamic_gap::Estimator* left_model, dynamic_gap::Estimator* right_model);
+            float gapSplinecheck(dynamic_gap::Gap & gap, dynamic_gap::Estimator*, dynamic_gap::Estimator*);
+            float generateCrossedGapTerminalPoints(float t, dynamic_gap::Gap & gap, 
+                                                    dynamic_gap::Estimator* left_model, 
+                                                    dynamic_gap::Estimator* right_model);
             void generateTerminalPoints(dynamic_gap::Gap & gap, 
-                                        double terminal_beta_left, double terminal_reciprocal_range_left, 
-                                        double terminal_beta_right, double terminal_reciprocal_range_right);
-            double indivGapFindCrossingPoint(dynamic_gap::Gap & gap, Eigen::Vector2f& gap_crossing_point, dynamic_gap::Estimator*, dynamic_gap::Estimator*);
+                                        float terminal_beta_left, float terminal_reciprocal_range_left, 
+                                        float terminal_beta_right, float terminal_reciprocal_range_right);
+            float indivGapFindCrossingPoint(dynamic_gap::Gap & gap, Eigen::Vector2f& gap_crossing_point, 
+                                            dynamic_gap::Estimator*, dynamic_gap::Estimator*);
 
             boost::shared_ptr<sensor_msgs::LaserScan const> msg;
-            double scan_angle_min_, scan_angle_increment_;
+            float scan_angle_min_, scan_angle_increment_;
             const DynamicGapConfig* cfg_;
             int num_of_scan;
             boost::mutex egolock;
